Many changes, see description
* Added use_normals parameter to Executor
* Added split_by_organism parameter to Executor
* Added only_human parameter to Executor
* Fixed problem in Executor with organism-dependent parameters
  (genome_ref, dbsnp...)
* Using a Runner class instead of the functool.partial
* Added --with-normal parameter
* When normals are used, Varscan & Mutect are run after everything else
  finished
* Changed filename convention "_" -> "."
* Dynamic parsing of sample names inside file list
* utils=rpackages.importr("utils") -> rutils to avoid conflicts with the
  module
* utils.get_params_from_filename does not use analysis parameter
* utils.get_params_from_filename correctly prints the filename which
  caused the error
* utils.get_sample_filenames can return a dict if the split parameter,
  just added, is true
* Added handling of tumor+normal for Mutect and Varscan
* Fixed utils.find_fastqs_by_organism
* Changed some parameters for VarScan with more pythonic names
* Added separate VarScan private runner for tumor+normal samples
* Xenograft performs a check on the files to be analyzed
* Xenograft now uses a sample-specific temporary directory to avoid
  filename conflicts

Solves some checks for card https://trello.com/c/rmuXj7Gw

+++ hatspil/xenograft.py
@@ -19,14 +19,13 @@ class Xenograft:
             "-T {analysis.config.xenome_threads} "
             "-P {analysis.config.xenome_index} --pairs")
         self.sample_base_out = os.path.join("REPORTS", self.analysis.sample)
+        self.skip_filenames = {}
 
     def chdir(self):
         os.chdir(self.fastq_dir)
 
     def xenome(self):
-        self.analysis.logger.info("Running xenome")
         self.chdir()
-
         input_filenames = [
             filename
             for pair in utils.find_fastqs_by_organism(
@@ -34,11 +33,35 @@ class Xenograft:
                 ".").values()
             for filename, _ in pair]
 
+        valid_filenames = []
+        for filename in input_filenames:
+            params = utils.get_params_from_filename(filename)
+            if params[2] == 60 and not params[8]:
+                valid_filenames.append(filename)
+            elif params[2] != 60:
+                organism = params[8]
+                if not organism:
+                    organism = "hg19"
+                if organism not in self.skip_filenames:
+                    self.skip_filenames[organism] = []
+                self.skip_filenames[organism].append(os.path.join(os.getcwd(), filename))
+
+        input_filenames = sorted(valid_filenames)
+
+        if len(input_filenames) == 0:
+            return
+
+        self.analysis.logger.info("Running xenome")
+        temp_dir = os.path.join(os.path.sep, "tmp", "%s.xenome" % self.analysis.sample)
+        if not os.path.exists(temp_dir):
+            os.mkdir(temp_dir)
+
         executor = Executor(self.analysis)
         executor(f(
             "{self.xenome_command} --graft-name hg19 --host-name mm10 "
             "--output-filename-prefix {self.analysis.sample} "
             "{{input_filename}} "
+            "--tmp-dir {temp_dir} "
             "> \"{self.sample_base_out}.xenome_summary.txt\""),
             input_filenames=input_filenames,
             input_function=lambda filenames: " ".join(["-i %s" % filename for filename in filenames]),
@@ -47,12 +70,18 @@ class Xenograft:
                                               for organism, index
                                               in itertools.product(["hg19", "mm10"],
                                                                    [1, 2])],
-            input_split_reads=False,
+            input_split_reads=False
         )
+        shutil.rmtree(temp_dir)
 
         self.analysis.logger.info("Finished xenome")
 
     def fix_fastq(self):
+        match = re.match(R"^[^-]+-[^-]+-(\d{2})", self.analysis.sample)
+        if not match or match.group(1) != "60":
+            self.analysis.last_operation_filenames = self.skip_filenames
+            return
+
         self.analysis.logger.info("Fixing xenome fastq")
         self.chdir()
         self.analysis.last_operation_filenames = {}
@@ -62,7 +91,7 @@ class Xenograft:
             match = re_fastq_filename.match(filename)
             organism = match.group(1)
             read_index = int(match.group(2))
-            out_filename = "%s_%s_R%d.fastq" % (self.analysis.sample, organism, read_index)
+            out_filename = "%s.%s.R%d.fastq" % (self.analysis.sample, organism, read_index)
             with open(filename, "r") as in_fd,\
                     open(out_filename, "w") as out_fd:
                 for line_index, line in enumerate(in_fd):
@@ -87,15 +116,17 @@ class Xenograft:
             for combo in itertools.product(["ambiguous", "both", "neither"],
                                            range(1, 3))]
         for filename in fastq_files + other_fastq:
-            os.unlink(filename)
+            if os.path.exists(filename):
+                os.unlink(filename)
 
+        self.analysis.last_operation_filenames.update(self.skip_filenames)
         self.analysis.logger.info("Finished fixing xenome fastq")
 
     def compress(self):
         self.analysis.logger.info("Compressing fastq files")
         self.chdir()
         fastq_files = [
-            "%s_R%d.fastq" % (self.analysis.sample, index + 1)
+            "%s\.R%d.fastq" % (self.analysis.sample, index + 1)
             for index in range(2)]
         for filename in fastq_files:
             compressed_filename = filename + ".gz"
